Move publish / schedule methods out of Publishable
These methods are no longer shared, so it doesn’t make sense to have
them in a module
