+++ lib/engines/content_block_manager/app/services/content_block_manager/create_edition_service.rb
@@ -1,7 +1,5 @@
 module ContentBlockManager
   class CreateEditionService
-    include Publishable
-
     def initialize(schema)
       @schema = schema
     end
